requirements.txt
fastapi
uvicorn[standard]
chromadb
pdfplumber
python-docx
pdf2image
pytesseract
Pillow
google-generativeai
python-dotenv
